# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentTODO API
  description: >
    <!-- @purpose -->AI 驱动的本地优先任务系统 RESTful API，支持可信网络免认证或可选 API Token，无缝对接 AI 助手和命令行技能系统。<!-- /purpose -->
  version: 1.0.0
  contact:
    name: AgentTODO
  license:
    name: MIT
    url: https://github.com/alanhsun/AgentTODO/blob/main/LICENSE
servers:
  - url: http://localhost:3300
    description: 本地开发

# API_TOKEN 未设置时允许匿名访问；设置后使用以下任一认证方式。
security:
  - bearerToken: []
  - basicAuth: []
  - {}
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        task_count:
          type: integer
...
    Subtask:
      type: object
      properties:
        id: { type: integer }
        task_id: { type: integer }
        title: { type: string }
        completed: { type: boolean }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
    TaskNote:
      type: object
      properties:
        id: { type: integer }
        task_id: { type: integer }
        content: { type: string }
        source: { type: string, enum: [user, ai] }
        created_at: { type: string, format: date-time }
...
paths:
...
  /api/tasks/{taskId}/subtasks:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: integer }
    get:
      summary: 获取子任务
      tags: [Subtasks]
      responses:
        '200':
          description: 子任务列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Subtask' }
    post:
      summary: 新增子任务
      tags: [Subtasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 255 }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subtask' }
  /api/tasks/{taskId}/subtasks/{id}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: integer }
      - name: id
        in: path
        required: true
        schema: { type: integer }
    put:
      summary: 更新子任务
      tags: [Subtasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 255 }
                completed: { type: boolean }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subtask' }
    delete:
      summary: 删除子任务
      tags: [Subtasks]
      responses:
        '200': { description: 删除成功 }
  /api/tasks/{taskId}/notes:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: integer }
    get:
      summary: 获取任务纪要
      tags: [Notes]
      responses:
        '200':
          description: 纪要列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TaskNote' }
    post:
      summary: 新增任务纪要
      tags: [Notes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                source: { type: string, enum: [user, ai], default: user }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskNote' }
  /api/tasks/{taskId}/notes/{id}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: integer }
      - name: id
        in: path
        required: true
        schema: { type: integer }
    delete:
      summary: 删除任务纪要
      tags: [Notes]
      responses:
        '200': { description: 删除成功 }
  /api/tags:
    get:
      summary: 获取标签列表
      tags: [Tags]
      responses:
        '200':
          description: 标签列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'

    post:
      summary: 创建标签
      tags: [Tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 50
                color:
                  type: string
                  pattern: '^#[0-9a-fA-F]{6}$'
                  default: '#6366f1'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '409':
          description: 标签名已存在
  /api/tags/{id}:
    put:
      summary: 更新标签
      tags: [Tags]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: 标签不存在

    delete:
      summary: 删除标签
      tags: [Tags]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '404':
          description: 标签不存在
...